extracted IFormAction to avoid circular reference
