+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-dynamic-form",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -30,53 +30,53 @@
     "@angular/platform-browser": "~13.3.11",
     "@angular/platform-browser-dynamic": "~13.3.11",
     "@angular/router": "~13.3.11",
-    "@sebgroup/ng-components": "^1.0.1",
-    "@storybook/addon-a11y": "^6.4.18",
+    "@sebgroup/ng-components": "~1.0.1",
+    "@storybook/addon-a11y": "~6.4.18",
     "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
+    "tslib": "~2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.9",
     "@angular/cli": "~13.3.9",
     "@angular/compiler-cli": "~13.3.11",
-    "@babel/core": "^7.14.6",
-    "@commitlint/cli": "^16.1.0",
-    "@commitlint/config-conventional": "^16.0.0",
-    "@compodoc/compodoc": "^1.1.11",
-    "@semantic-release/changelog": "^5.0.1",
-    "@storybook/addon-actions": "^6.4.18",
-    "@storybook/addon-essentials": "^6.4.18",
-    "@storybook/addon-links": "^6.4.18",
-    "@storybook/addons": "^6.4.18",
-    "@storybook/angular": "^6.4.18",
-    "@storybook/builder-webpack5": "^6.4.18",
-    "@storybook/manager-webpack5": "^6.4.18",
-    "@storybook/theming": "^6.4.18",
+    "@babel/core": "~7.14.6",
+    "@commitlint/cli": "~16.1.0",
+    "@commitlint/config-conventional": "~16.0.0",
+    "@compodoc/compodoc": "~1.1.11",
+    "@semantic-release/changelog": "~5.0.1",
+    "@storybook/addon-actions": "~6.4.18",
+    "@storybook/addon-essentials": "~6.4.18",
+    "@storybook/addon-links": "~6.4.18",
+    "@storybook/addons": "~6.4.18",
+    "@storybook/angular": "~6.4.18",
+    "@storybook/builder-webpack5": "~6.4.18",
+    "@storybook/manager-webpack5": "~6.4.18",
+    "@storybook/theming": "~6.4.18",
     "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
-    "babel-loader": "^8.2.2",
-    "codelyzer": "^6.0.0",
-    "gh-pages": "^3.2.2",
-    "git-cz": "^4.8.0",
-    "husky": "^7.0.4",
+    "@types/node": "~12.11.1",
+    "babel-loader": "~8.2.2",
+    "codelyzer": "~6.0.0",
+    "gh-pages": "~3.2.2",
+    "git-cz": "~4.8.0",
+    "husky": "~7.0.4",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.15",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^13.3.1",
-    "package-lock-sanitizer": "^1.0.1",
-    "prettier": "^2.5.1",
-    "pretty-quick": "^3.1.3",
+    "karma-jasmine-html-reporter": "~1.5.0",
+    "ng-packagr": "~13.3.1",
+    "package-lock-sanitizer": "~1.0.1",
+    "prettier": "~2.5.1",
+    "pretty-quick": "~3.1.3",
     "protractor": "~7.0.0",
-    "semantic-release": "^17.4.4",
+    "semantic-release": "~17.4.4",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
     "typescript": "~4.6.4",
-    "webpack": "^5.68.0"
+    "webpack": "~5.68.0"
   },
   "release": {
     "pkgRoot": "dist/dynamic-form",
