+++ projects/dynamic-form/src/lib/i-form-action.enum.ts
@@ -1,0 +1,6 @@
+export interface IFormAction {
+    hasAction: boolean;
+    label?: string;
+    theme?: string;
+    icon?: string;
+}
